requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0

# Embedding and similarity search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Configuration and utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Optional: for better performance with large datasets
pyarrow>=12.0.0